Added extension with patched PostgreSQL Java JDBC 4.2 (JRE 8+) driver version 9.4.1209 for PostgreSQL; the build differs from the original version provided by jdbc.postgresql.org insofar as the pre-8.4-behaviour in regards to autoGeneratedKeys is restored, so there's no more automatic appending of a RETURNING * clause to queries.
This feature of any official PostgreSQL JDBC driver starting with REL 8.4 breaks Connection.prepareStatement(String, int) for Lucee (and ACF), as any query at all automatically gets this RETURNING *  tacked on, eben if it's a SELECT statement or if there's already a RETURNING clause present, both of which cases lead to invalid SQL passed to the database server. Prior to 8.4, autoGeneratedKeys was not supported by the driver (so the flag was ignored), starting with 8.4, the driver would return all keys (not just sequences) by tacking on the RETURNING * to the query statement, when RETURN_GENERATED_KEYS was set (as is the case for any query from Lucee/ACF).

The patch just assumes autoGeneratedKeys to be always set to the NO_GENERATED_KEYS flag. This shouldn't make any difference to Lucee and CFML-code, but might require changes to sloppy SQL in other Java code that would run with this driver, as INSERT/UPDATE on PostgreSQL datasources would no longer automatically return keys.